app/requirements.txt
faiss_cpu==1.9.0.post1
langchain==0.3.13
langchain_core==0.3.28
langchain_groq==0.2.2
langgraph==0.2.60
numpy==2.2.1
python-dotenv==1.0.1
streamlit==1.41.1
transformers==4.47.1